requirements.txt
# requirements.txt (Final Version)
Flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.5.4
gunicorn==22.0.0